Add pop-up options menu
Now the user can configure options on a global basis from a pop-up menu
when clicking the WaspLine icon. Right now this includes gradient size and
the two colors used.

I'm aware that the color selector doesn't work on Firefox. It's not an
easy fix. In the future I want to add per-site configuration and a
variable number of colors.